+++ packages/jsii/test/symbol-identifiers.test.ts
@@ -7,7 +7,7 @@ test('Symbol map is generated', async () => {
         export * from './some/nested/file';
         export class Foo {
           public bar(){}
-        } 
+        }
       `,
       'some/nested/file.ts': `
         export interface Bar {
@@ -27,4 +27,68 @@ test('Symbol map is generated', async () => {
   expect(types['testpkg.Foo'].symbolId).toEqual('index:Foo');
   expect(types['testpkg.Bar'].symbolId).toEqual('some/nested/file:Bar');
   expect(types['testpkg.Baz'].symbolId).toEqual('some/nested/file:Baz');
+});
+
+test('Module declarations are included in symbolId', async () => {
+  const result = await compileJsiiForTest(
+    {
+      'index.ts': `
+        export class Foo {
+          constructor() {
+          }
+        }
+        export namespace Foo {
+          export class Bar {
+            public baz() {}
+          }
+        }
+      `,
+    },
+    undefined /* callback */,
+    { stripDeprecated: true },
+  );
+
+  const types = result.assembly.types ?? {};
+  expect(types['testpkg.Foo.Bar'].symbolId).toEqual('index:Foo.Bar');
+});
+
+test('Submodules also have symbol identifiers', async () => {
+  const result = await compileJsiiForTest(
+    {
+      'index.ts': `export * as submod from './submodule';`,
+      'submodule.ts': `
+        export class Foo {
+          constructor() {
+          }
+        }
+      `,
+    },
+    undefined /* callback */,
+    { stripDeprecated: true },
+  );
+
+  expect(result.assembly.submodules?.['testpkg.submod']?.symbolId).toEqual(
+    'submodule:',
+  );
+});
+
+test('Submodules also have symbol identifiers', async () => {
+  const result = await compileJsiiForTest(
+    {
+      'index.ts': `
+        export namespace cookie {
+          export class Foo {
+            constructor() {
+            }
+          }
+        }
+      `,
+    },
+    undefined /* callback */,
+    { stripDeprecated: true },
+  );
+
+  expect(result.assembly.submodules?.['testpkg.cookie']?.symbolId).toEqual(
+    'index:cookie',
+  );
 });